# api.yaml
swagger: '2.0'
info:
  title: API микросервиса общих справочников
  version: 1.0.0
paths:
...
  /dictionaries/branches#byCode:
    get:
      tags:
        - Branches
      description: Поиск филиала по коду
      operationId: getBranchByCode
      produces:
        - application/json
      parameters:
        - in: query
          name: code
          type: string
          description: Код филиала
          required: false
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/branch'
            description: Филиал с заданным кодом
        404:
          description: Not Found
  /dictionaries/branches/{id}:
    get:
      tags:
        - Branches
      description: Возвращает информацию о филиале
      operationId: getBranch
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор филиала
      responses:
        200:
          description: ОК
          schema:
            description: Информация о филиале
            $ref: '#/definitions/branch'
        404:
          description: Not Found


  #***channels***
...
  /dictionaries/channels/{id}:
    get:
      tags:
        - Channels
      description: Возвращает информацию о канале
      operationId: getChannel
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор канала
      responses:
        200:
          description: ОК
          schema:
            description: Информация о канале
            $ref: '#/definitions/channel'
        404:
          description: Not Found

  #***DivisionCodes***
...
  /dictionaries/cash-symbols:
    get:
      tags:
        - CashSymbols
      description: Список кассовых символов
      operationId: listCashSymbols
      x-spring-pageable: true
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/cash-symbol'
            description: Постраничный вывод результатов поиска
        403:
          description: Forbidden
  /dictionaries/cash-symbols#byCode:
    get:
      tags:
        - CashSymbols
      description: Возвращает информацию о кассовом символе по коду
      operationId: getCashSymbolByCode
      produces:
        - application/json
      parameters:
        - in: query
          name: code
          type: integer
          format: int64
          required: true
          description: Код кассового символа
      responses:
        200:
          description: ОК
          schema:
            description: Информация о кассовом символе
            $ref: '#/definitions/cash-symbol'
        403:
          description: Forbidden
        404:
          description: Not Found
  /dictionaries/cash-symbols/{id}:
    get:
      tags:
        - CashSymbols
      description: Возвращает информацию о кассовом символе
      operationId: getCashSymbol
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор кассового символа
      responses:
        200:
          description: ОК
          schema:
            description: Информация о кассовом символе
            $ref: '#/definitions/cash-symbol'
        403:
          description: Forbidden
        404:
          description: Not Found
...
  /dictionaries/blockages/check:
    post:
      tags:
        - Blockages
      description: Проверка паспортных данных по справочнику блокировок/заморозок
      operationId: checkBlockages
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: checkBlockagesRq
          description: Тело запроса проверки паспортных данных по справочнику блокировок/заморозок
          required: true
          schema:
            $ref: '#/definitions/checkBlockagesRq'
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/checkBlockagesRs'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /dictionaries/blockages/update:
    get:
      tags:
        - Blockages
      description: Запрос обновления списка блокировок/заморозок
      operationId: updateBlockageList
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/updateListStatus'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
    post:
      tags:
        - Blockages
      operationId: uploadBlockages
      parameters:
        - in: formData
          name: content
          type: file
          description: XML файл-справочник заморозок/блокировок
      responses:
        200:
          description: ОК
        403:
          description: Forbidden
        404:
          description: Not Found
        500:
          description: Internal server error
  /dictionaries/blockages/status:
    get:
      tags:
        - Blockages
      description: Получение текущего статуса процесса обновления справочника блокировок/заморозок
      operationId: getBlockageUpdateStatus
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/updateListStatus'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
...
definitions:
...
  branch:
    type: object
    description: Филиал
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      code:
        type: string
        description: Код филиала
      simpleName:
        type: string
        description: Краткое наименование
      officialName:
        type: string
        description: Официальное наименование
  channel:
    type: object
    description: Канал
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      shortName:
        type: string
        description: Краткое наименование
      fullName:
        type: string
        description: Полное наименование
...
  operationType:
    type: string
    description: Тип операции
    enum:
      - income
      - expenditure
  cash-symbol:
    description: Кассовый символ
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      code:
        type: integer
        format: int64
        description: Код
      name:
        type: string
        description: Наименование
      operationType:
        $ref: '#/definitions/operationType'
        description: Применение
      purposeOfPayment:
        type: string
        description: Назначение платежа
  updateListStatus:
    type: object
    properties:
      status:
        type: string
...
  personnelData:
    type: object
    properties:
      id:
        type: string
        description: Уникальный (в рамках запроса) идентификатор для обратной взаимосвязи
      lastName:
        type: string
        description: Фамилия
      firstName:
        type: string
        description: Имя
      middleName:
        type: string
        description: Отчество
      birthDate:
        type: string
        format: date
        description: Дата рождения
      passportSeries:
        type: string
        description: Серия паспорта
      passportNumber:
        type: string
        description: Номер паспорта
...
  checkBlockagesRq:
    type: object
    properties:
      citizens:
        type: array
        items:
          $ref: '#/definitions/personnelData'
...
  checkBlockagesResult:
    type: object
    properties:
      id:
        type: string
        description: Идентификатор записи
      isBlock:
        type: boolean
        description: Результат проверки
      additionalInf:
        type: string
        description: Подробности
...
  checkBlockagesRs:
    type: object
    properties:
      updateId:
        type: string
        description: Идентификатор записи содержащей информацию о файле-справочнике
      citizens:
        type: array
        items:
          $ref: '#/definitions/checkBlockagesResult'
...