# common-dict-app/src/main/java/ru/softlab/efr/common/dict/controllers/BranchesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package ru.softlab.efr.common.dict.controllers;

import ru.softlab.efr.common.dict.exchange.model.Branch;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.validation.BindingResult;
import org.springframework.data.web.PageableDefault;
import org.springframework.data.domain.*;

import java.util.*;

import org.springframework.validation.annotation.Validated;
import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * the Branches API
 */
public interface BranchesApi {

    /**
     * Возвращает информацию о филиале 
     * @param id Идентификатор филиала  
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/branches/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Branch> getBranch(@PathVariable("id") Long id) throws Exception;



    /**
     * Поиск филиала по коду 
     * @param code Код филиала  
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/branches#byCode",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Branch> getBranchByCode( @Valid @RequestParam(value = "code", required = false) String code) throws Exception;



    /**
     * Список филиалов   
     * @param pageable параметры пагинации 
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/dictionaries/branches",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Page<Branch>> listBranches(@PageableDefault(value = 50) Pageable pageable) throws Exception;


}